fb_bot/requirements.txt
aenum==2.2.6
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
fbchat==1.9.7
idna==3.3
paho-mqtt==1.6.1
python-dotenv==0.19.2
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.9
